Updated to remove all functions in computeFunctions. Now dependence is explicit. All functions are in respective py files.
