+++ src/go_melt/core/heat_source_functions.py
@@ -1,13 +1,13 @@
 from functools import partial
 import jax
 import jax.numpy as jnp
-from .computeFunctions import *
 from jax.numpy import multiply
 from go_melt.utils.gaussian_quadrature_functions import (
     getQuadratureCoords,
     computeQuad3dFemShapeFunctions_jax,
 )
-from go_melt.utils.helper_functions import convert2XYZ
+from go_melt.utils.helper_functions import convert2XYZ, bincount
+from .mesh_functions import getSampleCoords
 
 
 @partial(jax.jit, static_argnames=["ne_nn"])
